this commit improves readability and performance
Remove TERMINATE statements. Calculating the matrix elements does not seem to
be any kind of problem, so we remove the code.

Reduce calls to logadd_ms. We just one to add two numbers (given in log
representation), so we may just use logadd_s. Also, we don't insert the Mie
coefficients into the sum, so we can reuse results and decrease the calls to
logadd_*. Also, this helps to improve readability.