Check matrix elements, Tr approximation for Drude
This commit consists of two parts:
We introduce a new flag check_elems. If set to true, all matrix elements of the
round-trip matrix will be initialized with NAN before calculation. After
calculation it is checked if every matrix element is finite.

Trace approxumation for Drude:
If the trace of the round trip is small, we don't need to calculate all matrix
elements. This worked for perfect reflectors up to now. With this commit this
will also work for Drude metals.
Also, this commit fixes a bug, because we didn't compare the absolute value of
the trace with the threshold.