Remove support for Romberg integration
This commit removes support for Romberg integration for following reasons:
 * The implementation of Romberg integration is slow.
 * The quality and documentation of the code was poor.
 * No accurate error estimation. (I'm not sure if error estimation really
   worked.)

At the moment Drude support is only supported using fixed order Gauss-Laguerre
integration. This method is not reliable for lmax becoming large, i.e. for
small separations L.

The idea is to add support for Drude or arbitrary non-magnetic metals in the
future using an adaptive Gauss-Kronrod integration scheme.