Support metals with arbitrary epsilon(i*xi)
This commit provides the infrastructure to support arbitrary metals. The
behavior of the reflectors are determined by the dielectric function. The user
can give a callback to an arbitrary function that calculates (to avoid loss of
significance) epsilon(i*xi)-1. The user can also pass a pointer to arbitrary
data that will be given to the callback whenever called.