Fix contribution of xi=0 for plasma model
The contribution of the zero-th Matsubara frequency in the plasma model was
wrong. The program gave the wrong parameter to the function
casimir_logdetD0_plasma which expects the scaled plasma frequency
omegap*(L+R)/c instead of the plasma frequency.

This bug only occured when casimir was started with a value for omegap and
without a value for gamma.

No publication is affected by this bug.