paragraph removed
I do not think that the statement about the factor of 2 is correct
as it was. As the discussion is not strictly needed here, I removed
it, but of course it could be replaced by a corrected discussion of
the differences between Drude and plasma. But may be this would
require to go too much into the details.